UIFC-338-use-functional-components-instead-of-class-components improve FindSource
